Adicionado rotação de reta e de polígono :D

+++ listaRetas.h
@@ -18,7 +18,8 @@ void ListaRetasInserirFim(ListaRetas* lista, Reta r);
 void desenharRetas(ListaRetas* lista);
 int codigo(double x, double y, double xmin, double xmax, double ymin, double ymax);
 int pickReta(double x0, double y0, double x1, double y1, double mx, double my, double t);
-void selecionarRetaMaisProxima(ListaRetas* lista, double x, double y);
+//void selecionarRetaMaisProxima(ListaRetas* lista, double x, double y);
+void selecionarRetaMaisProxima(ListaRetas* lista, int x, int y) ;
 void deletarRetasSelecionadas(ListaRetas* lista);
 
 #endif
